[REFACTOR/#52] TradeType core > enum으로 변경
