+++ app/src/main/java/com/napzak/market/presentation/explore/component/TradeTypeTab.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.napzak.market.core.common.extension.noRippleClickable
 import com.napzak.market.core.designsystem.theme.NapzakMarketTheme
-import com.napzak.market.presentation.explore.type.TradeType
+import com.napzak.market.core.type.TradeType
 
 @Composable
 fun TradeTypeTab(
