+++ app/src/main/java/com/napzak/market/presentation/explore/state/ExploreUiState.kt
@@ -3,10 +3,10 @@ package com.napzak.market.presentation.explore.state
 import androidx.compose.runtime.Immutable
 import com.napzak.market.core.common.state.UiState
 import com.napzak.market.core.type.SortType
+import com.napzak.market.core.type.TradeType
 import com.napzak.market.domain.explore.model.ProductItem
 import com.napzak.market.domain.genre.model.Genre
 import com.napzak.market.presentation.explore.type.ExploreScreenType
-import com.napzak.market.presentation.explore.type.TradeType
 
 @Immutable
 data class ExploreUiState(
